abstract class content gemaakt
