+++ src/main/java/nl/bo/bacpatternpoint/models/Pattern.java
@@ -11,18 +11,9 @@ import java.util.List;
 
 @Entity
 @Table(name = "patterns")
-public class Pattern {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    @NotNull(message = "Titel is verplicht")
-    @Size(min = 5, max = 50, message = "Titel moet tussen 5 en 50 karakters zijn")
-    private String title;
+public class Pattern extends Content{
     @NotNull(message = "Niveau is verplicht")
     private String level;
-    @NotNull(message = "Beschrijving is verplicht")
-    @Size(min = 5, max = 300, message = "Beschrijving moet tussen 5 en 300 karakters zijn")
-    private String description;
     @Positive(message = "Haaknaald moet groter dan 0mm zijn")
     private double hookSize;
     @Min(value = 1, message = "Aantal gram wol moet minimaal 1 zijn")
@@ -35,7 +26,6 @@ public class Pattern {
     private double length;
     @Positive(message = "Breedte moet groter dan 0cm zijn")
     private double width;
-    private boolean isDraft;
     @OneToOne(cascade = CascadeType.ALL)
     private Image image;
     @OneToMany(mappedBy = "pattern", cascade = CascadeType.ALL, fetch = FetchType.EAGER)
@@ -57,14 +47,6 @@ public class Pattern {
 
     public void setPost(Post post) {
         this.post = post;
-    }
-
-    public boolean isDraft() {
-        return isDraft;
-    }
-
-    public void setDraft(boolean draft) {
-        isDraft = draft;
     }
 
     public User getUser() {
@@ -171,36 +153,12 @@ public class Pattern {
         this.hookSize = hookSize;
     }
 
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
     public String getLevel() {
         return level;
     }
 
     public void setLevel(String level) {
         this.level = level;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
     }
 }
