+++ src/main/java/nl/bo/bacpatternpoint/models/Post.java
@@ -2,32 +2,21 @@ package nl.bo.bacpatternpoint.models;
 
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Size;
 
 import java.util.List;
 
 @Entity
 @Table(name = "posts")
-public class Post {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    @NotNull(message = "Titel is verplicht")
-    @Size(min = 5, max = 50, message = "Titel moet tussen 5 en 50 karakters zijn")
-    private String title;
+public class Post extends Content {
     @NotNull(message = "Categorie is verplicht")
     private String category;
-    @NotNull(message = "Beschrijving is verplicht")
-    @Size(min = 5, max = 300, message = "Beschrijving moet tussen 5 en 300 karakters zijn")
-    private String description;
-    private int likes;
-    private boolean isDraft;
-    @OneToOne(cascade = CascadeType.ALL)
-    private Image image;
+
     @OneToMany(mappedBy = "post", cascade = CascadeType.ALL)
     private List<Comment> comments;
+
     @OneToMany(mappedBy = "post")
     private List<Pattern> patterns;
+
     @ManyToOne
     @JoinColumn(name = "user_id", referencedColumnName = "id")
     private User user;
@@ -56,63 +45,11 @@ public class Post {
         this.comments = comments;
     }
 
-    public Image getImage() {
-        return image;
-    }
-
-    public void setImage(Image image) {
-        this.image = image;
-    }
-
-    public long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public void setId(long id) {
-        this.id = id;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
     public String getCategory() {
         return category;
     }
 
     public void setCategory(String category) {
         this.category = category;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public int getLikes() {
-        return likes;
-    }
-
-    public void setLikes(int likes) {
-        this.likes = likes;
-    }
-
-    public boolean isDraft() {
-        return isDraft;
-    }
-
-    public void setDraft(boolean draft) {
-        isDraft = draft;
     }
 }
